move status client to new package
